[TMP] Completed append processing and reporting of samples to the upper layer
Now the buffered ranges start to appear and the appends complete. I've added a new
state machine (AppendStage + setAppendStage()) to control the different stages of
the append operation. This simplifies the management of the several timeouts in
AppendPipeline.

The majority of the tests 1..26 pass (except 20. DASHLatency), but there's a
problem in 23. DurationAfterAppendAudio. I think it's caused by a problem in the
timestamps generation: the same audio data is appended twice in the test, and
while the buffered ranges should overlap, the second time the buffered ranges
extend beyond what they were in the first time. Might be a problem in the demuxer,
but it's going to cause trouble in the seek in the future.