[TMP] Partial setup of the multiple streams in the playback pipeline
Streams in the playback pipeline are setup in several stages: In the initial
stage, only the appsrc elements are created (no caps, we still don't know the
kind of track that stream will become). In the "track detected" stage the
append pipeline has been completely setup, so its demuxer has the caps we
need to setup the stream in the playback pipeline. Caps are set in the
playback pipeline appsrc and the rest of the element chain is created (this
is still incomplete).

For practical reasons, I've dropped the support of multiple Streams per
Source in WebKitMediaSrc. By now, both structures (Source, Stream) are kept,
but the list of Streams has been demoted to a single pointer. Source is
created when the SourceBufferPrivateGStreamer is added to the
PlaybackPipeline. This creates the unconfigured appsrc element for that
Source-Stream. Stream is created when the track info is detected in the
append pipeline and attached to the playback pipeline.

I have to continue working on PlaybackPipeline::attachTrack().